Add support for multiple rotation angles
This commit adds support for specifying multiple rotation angles to be
used by the script. If fewer then the required number of angles are
specified, the last one is reused for any remaining geometries. There is
not any real error checking nor validation for this feature, nor really
any of the supported command line arguments.